Module google_docs1::api

Structs

AutoText
A ParagraphElement representing a spot in the text that is dynamically replaced with content that can change over time, like a page number.

Background
Represents the background of a document.

BackgroundSuggestionState
A mask that indicates which of the fields on the base Background have been changed in this suggestion. For any field set to true, the Backgound has a new suggested value.

BatchUpdateDocumentRequest
Request message for BatchUpdateDocument.

BatchUpdateDocumentResponse
Response message from a BatchUpdateDocument request.

Body
The document body. The body typically contains the full document contents except for headers, footers and footnotes.

Bullet
Describes the bullet of a paragraph.

BulletSuggestionState
A mask that indicates which of the fields on the base Bullet have been changed in this suggestion. For any field set to true, there is a new suggested value.

Color
A solid color.

ColumnBreak
A ParagraphElement representing a column break. A column break makes the subsequent text start at the top of the next column.

CreateFooterRequest
Creates a Footer. The new footer is applied to the SectionStyle at the location of the SectionBreak if specificed, otherwise it is applied to the DocumentStyle. If a footer of the specified type already exists, a 400 bad request error is returned.

CreateFooterResponse
The result of creating a footer.

CreateFootnoteRequest
Creates a Footnote segment and inserts a new FootnoteReference to it at the given location. The new Footnote segment will contain a space followed by a newline character.

CreateFootnoteResponse
The result of creating a footnote.

CreateHeaderRequest
Creates a Header. The new header is applied to the SectionStyle at the location of the SectionBreak if specificed, otherwise it is applied to the DocumentStyle. If a header of the specified type already exists, a 400 bad request error is returned.

CreateHeaderResponse
The result of creating a header.

CreateNamedRangeRequest
Creates a NamedRange referencing the given range.

CreateNamedRangeResponse
The result of creating a named range.

CreateParagraphBulletsRequest
Creates bullets for all of the paragraphs that overlap with the given range. The nesting level of each paragraph will be determined by counting leading tabs in front of each paragraph. To avoid excess space between the bullet and the corresponding paragraph, these leading tabs are removed by this request. This may change the indices of parts of the text. If the paragraph immediately before paragraphs being updated is in a list with a matching preset, the paragraphs being updated are added to that preceding list.

CropProperties
The crop properties of an image. The crop rectangle is represented using fractional offsets from the original content's four edges. - If the offset is in the interval (0, 1), the corresponding edge of crop rectangle is positioned inside of the image's original bounding rectangle. - If the offset is negative or greater than 1, the corresponding edge of crop rectangle is positioned outside of the image's original bounding rectangle. - If all offsets and rotation angle are 0, the image is not cropped.

CropPropertiesSuggestionState
A mask that indicates which of the fields on the base CropProperties have been changed in this suggestion. For any field set to true, there is a new suggested value.

DeleteContentRangeRequest
Deletes content from the document.

DeleteFooterRequest
Deletes a Footer from the document.

DeleteHeaderRequest
Deletes a Header from the document.

DeleteNamedRangeRequest
Deletes a NamedRange.

DeleteParagraphBulletsRequest
Deletes bullets from all of the paragraphs that overlap with the given range. The nesting level of each paragraph will be visually preserved by adding indent to the start of the corresponding paragraph.

DeletePositionedObjectRequest
Deletes a PositionedObject from the document.

DeleteTableColumnRequest
Deletes a column from a table.

DeleteTableRowRequest
Deletes a row from a table.

Dimension
A magnitude in a single direction in the specified units.

Docs
Central instance to access all Docs related resource activities

Document
A Google Docs document.

DocumentBatchUpdateCall
Applies one or more updates to the document. Each request is validated before being applied. If any request is not valid, then the entire request will fail and nothing will be applied. Some requests have replies to give you some information about how they are applied. Other requests do not need to return information; these each return an empty reply. The order of replies matches that of the requests. For example, suppose you call batchUpdate with four updates, and only the third one returns information. The response would have two empty replies, the reply to the third request, and another empty reply, in that order. Because other users may be editing the document, the document might not exactly reflect your changes: your changes may be altered with respect to collaborator changes. If there are no collaborators, the document should reflect your changes. In any case, the updates in your request are guaranteed to be applied together atomically.

DocumentCreateCall
Creates a blank document using the title given in the request. Other fields in the request, including any provided content, are ignored. Returns the created document.

DocumentGetCall
Gets the latest version of the specified document.

DocumentMethods
A builder providing access to all methods supported on document resources. It is not used directly, but through the Docs hub.

DocumentStyle
The style of the document.

DocumentStyleSuggestionState
A mask that indicates which of the fields on the base DocumentStyle have been changed in this suggestion. For any field set to true, there is a new suggested value.

EmbeddedDrawingProperties
The properties of an embedded drawing.

EmbeddedDrawingPropertiesSuggestionState
A mask that indicates which of the fields on the base EmbeddedDrawingProperties have been changed in this suggestion. For any field set to true, there is a new suggested value.

EmbeddedObject
An embedded object in the document.

EmbeddedObjectBorder
A border around an EmbeddedObject.

EmbeddedObjectBorderSuggestionState
A mask that indicates which of the fields on the base EmbeddedObjectBorder have been changed in this suggestion. For any field set to true, there is a new suggested value.

EmbeddedObjectSuggestionState
A mask that indicates which of the fields on the base EmbeddedObject have been changed in this suggestion. For any field set to true, there is a new suggested value.

EndOfSegmentLocation
Location at the end of a body, header, footer or footnote. The location is immediately before the last newline in the document segment.

Equation
A ParagraphElement representing an equation.

Footer
A document footer.

Footnote
A document footnote.

FootnoteReference
A ParagraphElement representing a footnote reference. A footnote reference is the inline content rendered with a number and is used to identify the footnote.

Header
A document header.

HorizontalRule
A ParagraphElement representing a horizontal line.

ImageProperties
The properties of an image.

ImagePropertiesSuggestionState
A mask that indicates which of the fields on the base ImageProperties have been changed in this suggestion. For any field set to true, there is a new suggested value.

InlineObject
An object that appears inline with text. An InlineObject contains an EmbeddedObject such as an image.

InlineObjectElement
A ParagraphElement that contains an InlineObject.

InlineObjectProperties
Properties of an InlineObject.

InlineObjectPropertiesSuggestionState
A mask that indicates which of the fields on the base InlineObjectProperties have been changed in this suggestion. For any field set to true, there is a new suggested value.

InsertInlineImageRequest
Inserts an InlineObject containing an image at the given location.

InsertInlineImageResponse
The result of inserting an inline image.

InsertInlineSheetsChartResponse
The result of inserting an embedded Google Sheets chart.

InsertPageBreakRequest
Inserts a page break followed by a newline at the specified location.

InsertSectionBreakRequest
Inserts a section break at the given location. A newline character will be inserted before the section break.

InsertTableColumnRequest
Inserts an empty column into a table.

InsertTableRequest
Inserts a table at the specified location. A newline character will be inserted before the inserted table.

InsertTableRowRequest
Inserts an empty row into a table.

InsertTextRequest
Inserts text at the specified location.

Link
A reference to another portion of a document or an external URL resource.

LinkedContentReference
A reference to the external linked source content.

LinkedContentReferenceSuggestionState
A mask that indicates which of the fields on the base LinkedContentReference have been changed in this suggestion. For any field set to true, there is a new suggested value.

List
A List represents the list attributes for a group of paragraphs that all belong to the same list. A paragraph that is part of a list has a reference to the list's ID in its bullet.

ListProperties
The properties of a list which describe the look and feel of bullets belonging to paragraphs associated with a list.

ListPropertiesSuggestionState
A mask that indicates which of the fields on the base ListProperties have been changed in this suggestion. For any field set to true, there is a new suggested value.

Location
A particular location in the document.

MergeTableCellsRequest
Merges cells in a Table.

NamedRange
A collection of Ranges with the same named range ID. Named ranges allow developers to associate parts of a document with an arbitrary user-defined label so their contents can be programmatically read or edited at a later time. A document can contain multiple named ranges with the same name, but every named range has a unique ID. A named range is created with a single Range, and content inserted inside a named range generally expands that range. However, certain document changes can cause the range to be split into multiple ranges. Named ranges are not private. All applications and collaborators that have access to the document can see its named ranges.

NamedRanges
A collection of all the NamedRanges in the document that share a given name.

NamedStyle
A named style. Paragraphs in the document can inherit their TextStyle and ParagraphStyle from this named style when they have the same named style type.

NamedStyleSuggestionState
A suggestion state of a NamedStyle message.

NamedStyles
The named styles. Paragraphs in the document can inherit their TextStyle and ParagraphStyle from these named styles.

NamedStylesSuggestionState
The suggestion state of a NamedStyles message.

NestingLevel
Contains properties describing the look and feel of a list bullet at a given level of nesting.

NestingLevelSuggestionState
A mask that indicates which of the fields on the base NestingLevel have been changed in this suggestion. For any field set to true, there is a new suggested value.

ObjectReferences
A collection of object IDs.

OptionalColor
A color that can either be fully opaque or fully transparent.

PageBreak
A ParagraphElement representing a page break. A page break makes the subsequent text start at the top of the next page.

Paragraph
A StructuralElement representing a paragraph. A paragraph is a range of content that is terminated with a newline character.

ParagraphBorder
A border around a paragraph.

ParagraphElement
A ParagraphElement describes content within a Paragraph.

ParagraphStyle
Styles that apply to a whole paragraph. Inherited paragraph styles are represented as unset fields in this message. A paragraph style's parent depends on where the paragraph style is defined: * The ParagraphStyle on a Paragraph inherits from the paragraph's corresponding named style type. * The ParagraphStyle on a named style inherits from the normal text named style. * The ParagraphStyle of the normal text named style inherits from the default paragraph style in the Docs editor. * The ParagraphStyle on a Paragraph element that is contained in a table may inherit its paragraph style from the table style. If the paragraph style does not inherit from a parent, unsetting fields will revert the style to a value matching the defaults in the Docs editor.

ParagraphStyleSuggestionState
A mask that indicates which of the fields on the base ParagraphStyle have been changed in this suggestion. For any field set to true, there is a new suggested value.

Person
A person or email address mentioned in a document. These mentions behave as a single, immutable element containing the person's name or email address.

PersonProperties
Properties specific to a linked Person.

PositionedObject
An object that is tethered to a Paragraph and positioned relative to the beginning of the paragraph. A PositionedObject contains an EmbeddedObject such as an image.

PositionedObjectPositioning
The positioning of a PositionedObject. The positioned object is positioned relative to the beginning of the Paragraph it is tethered to.

PositionedObjectPositioningSuggestionState
A mask that indicates which of the fields on the base PositionedObjectPositioning have been changed in this suggestion. For any field set to true, there is a new suggested value.

PositionedObjectProperties
Properties of a PositionedObject.

PositionedObjectPropertiesSuggestionState
A mask that indicates which of the fields on the base PositionedObjectProperties have been changed in this suggestion. For any field set to true, there is a new suggested value.

Range
Specifies a contiguous range of text.

ReplaceAllTextRequest
Replaces all instances of text matching a criteria with replace text.

ReplaceAllTextResponse
The result of replacing text.

ReplaceImageRequest
Replaces an existing image with a new image. Replacing an image removes some image effects from the existing image in order to mirror the behavior of the Docs editor.

ReplaceNamedRangeContentRequest
Replaces the contents of the specified NamedRange or NamedRanges with the given replacement content. Note that an individual NamedRange may consist of multiple discontinuous ranges. In this case, only the content in the first range will be replaced. The other ranges and their content will be deleted. In cases where replacing or deleting any ranges would result in an invalid document structure, a 400 bad request error is returned.

Request
A single update to apply to a document.

Response
A single response from an update.

RgbColor
An RGB color.

SectionBreak
A StructuralElement representing a section break. A section is a range of content which has the same SectionStyle. A section break represents the start of a new section, and the section style applies to the section after the section break. The document body always begins with a section break.

SectionColumnProperties
Properties that apply to a section's column.

SectionStyle
The styling that applies to a section.

Shading
The shading of a paragraph.

ShadingSuggestionState
A mask that indicates which of the fields on the base Shading have been changed in this suggested change. For any field set to true, there is a new suggested value.

SheetsChartReference
A reference to a linked chart embedded from Google Sheets.

SheetsChartReferenceSuggestionState
A mask that indicates which of the fields on the base SheetsChartReference have been changed in this suggestion. For any field set to true, there is a new suggested value.

Size
A width and height.

SizeSuggestionState
A mask that indicates which of the fields on the base Size have been changed in this suggestion. For any field set to true, the Size has a new suggested value.

StructuralElement
A StructuralElement describes content that provides structure to the document.

SubstringMatchCriteria
A criteria that matches a specific string of text in the document.

SuggestedBullet
A suggested change to a Bullet.

SuggestedDocumentStyle
A suggested change to the DocumentStyle.

SuggestedInlineObjectProperties
A suggested change to InlineObjectProperties.

SuggestedListProperties
A suggested change to ListProperties.

SuggestedNamedStyles
A suggested change to the NamedStyles.

SuggestedParagraphStyle
A suggested change to a ParagraphStyle.

SuggestedPositionedObjectProperties
A suggested change to PositionedObjectProperties.

SuggestedTableCellStyle
A suggested change to a TableCellStyle.

SuggestedTableRowStyle
A suggested change to a TableRowStyle.

SuggestedTextStyle
A suggested change to a TextStyle.

TabStop
A tab stop within a paragraph.

Table
A StructuralElement representing a table.

TableCell
The contents and style of a cell in a Table.

TableCellBorder
A border around a table cell. Table cell borders cannot be transparent. To hide a table cell border, make its width 0.

TableCellLocation
Location of a single cell within a table.

TableCellStyle
The style of a TableCell. Inherited table cell styles are represented as unset fields in this message. A table cell style can inherit from the table's style.

TableCellStyleSuggestionState
A mask that indicates which of the fields on the base TableCellStyle have been changed in this suggestion. For any field set to true, there is a new suggested value.

TableColumnProperties
The properties of a column in a table.

TableOfContents
A StructuralElement representing a table of contents.

TableRange
A table range represents a reference to a subset of a table. It's important to note that the cells specified by a table range do not necessarily form a rectangle. For example, let's say we have a 3 x 3 table where all the cells of the last row are merged together. The table looks like this: [ ] A table range with table cell location = (table_start_location, row = 0, column = 0), row span = 3 and column span = 2 specifies the following cells: x x [ x x x ]

TableRow
The contents and style of a row in a Table.

TableRowStyle
Styles that apply to a table row.

TableRowStyleSuggestionState
A mask that indicates which of the fields on the base TableRowStyle have been changed in this suggestion. For any field set to true, there is a new suggested value.

TableStyle
Styles that apply to a table.

TextRun
A ParagraphElement that represents a run of text that all has the same styling.

TextStyle
Represents the styling that can be applied to text. Inherited text styles are represented as unset fields in this message. A text style's parent depends on where the text style is defined: * The TextStyle of text in a Paragraph inherits from the paragraph's corresponding named style type. * The TextStyle on a named style inherits from the normal text named style. * The TextStyle of the normal text named style inherits from the default text style in the Docs editor. * The TextStyle on a Paragraph element that is contained in a table may inherit its text style from the table style. If the text style does not inherit from a parent, unsetting fields will revert the style to a value matching the defaults in the Docs editor.

TextStyleSuggestionState
A mask that indicates which of the fields on the base TextStyle have been changed in this suggestion. For any field set to true, there is a new suggested value.

UnmergeTableCellsRequest
Unmerges cells in a Table.

UpdateDocumentStyleRequest
Updates the DocumentStyle.

UpdateParagraphStyleRequest
Update the styling of all paragraphs that overlap with the given range.

UpdateSectionStyleRequest
Updates the SectionStyle.

UpdateTableCellStyleRequest
Updates the style of a range of table cells.

UpdateTableColumnPropertiesRequest
Updates the TableColumnProperties of columns in a table.

UpdateTableRowStyleRequest
Updates the TableRowStyle of rows in a table.

UpdateTextStyleRequest
Update the styling of text.

WeightedFontFamily
Represents a font family and weight of text.

WriteControl
Provides control over how write requests are executed.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.