Struct google_docs1::TextStyle

pub struct TextStyle {
    pub foreground_color: Option<OptionalColor>,
    pub bold: Option<bool>,
    pub baseline_offset: Option<String>,
    pub strikethrough: Option<bool>,
    pub weighted_font_family: Option<WeightedFontFamily>,
    pub small_caps: Option<bool>,
    pub font_size: Option<Dimension>,
    pub italic: Option<bool>,
    pub link: Option<Link>,
    pub underline: Option<bool>,
    pub background_color: Option<OptionalColor>,
}

Represents the styling that can be applied to text.

Inherited text styles are represented as unset fields in this message. A text style's parent depends on where the text style is defined:

The TextStyle of text in a Paragraph inherits from the paragraph's corresponding named style type.

The TextStyle on a named style inherits from the normal text named style.

The TextStyle of the normal text named style inherits from the default text style in the Docs editor.

The TextStyle on a Paragraph element that is contained in a table may inherit its text style from the table style.

If the text style does not inherit from a parent, unsetting fields will revert the style to a value matching the defaults in the Docs editor.

This type is not used in any activity, and only used as part of another schema.

Fields
foreground_color: Option<OptionalColor>

The foreground color of the text. If set, the color is either an RGB color or transparent, depending on the color field.

bold: Option<bool>

Whether or not the text is rendered as bold.

baseline_offset: Option<String>

The text's vertical offset from its normal position.

Text with SUPERSCRIPT or SUBSCRIPT baseline offsets is automatically rendered in a smaller font size, computed based on the font_size field. The font_size itself is not affected by changes in this field.

strikethrough: Option<bool>

Whether or not the text is struck through.

weighted_font_family: Option<WeightedFontFamily>

The font family and rendered weight of the text.

If an update request specifies values for both weighted_font_family and bold, the weighted_font_family is applied first, then bold.

If weighted_font_family#weight is not set, it defaults to 400.

If weighted_font_family is set, then weighted_font_family#font_family must also be set with a non-empty value. Otherwise, a 400 bad request error is returned.

small_caps: Option<bool>

Whether or not the text is in small capital letters.

font_size: Option<Dimension>

The size of the text's font.

italic: Option<bool>

Whether or not the text is italicized.

link: Option<Link>

The hyperlink destination of the text. If unset, there is no link. Links are not inherited from parent text.

Changing the link in an update request causes some other changes to the text style of the range:

When setting a link, the text foreground color will be updated to the default link color and the text will be underlined. If these fields are modified in the same request, those values will be used instead of the link defaults.

Setting a link on a text range that overlaps with an existing link will also update the existing link to point to the new URL.

Links are not settable on newline characters. As a result, setting a link on a text range that crosses a paragraph boundary, such as "ABC\n123", will separate the newline character(s) into their own text runs. The link will be applied separately to the runs before and after the newline.

Removing a link will update the text style of the range to match the style of the preceding text (or the default text styles if the preceding text is another link) unless different styles are being set in the same request.

underline: Option<bool>

Whether or not the text is underlined.

background_color: Option<OptionalColor>

The background color of the text. If set, the color is either an RGB color or transparent, depending on the color field.